api/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord>=2.3.2",
    "fastapi[standard]>=0.115.13",
    "ffmpeg>=1.4",
    "mcp>=1.10.0",
    "mutagen>=1.47.0",
    "pydantic>=2.11.7",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.1",
    "websockets>=15.0.1",
    "yt-dlp>=2025.6.9",
]
